chore(load-balancers): update generated types and methods

+++ src/resources/load-balancers/pools/pools.ts
@@ -1,6 +1,7 @@
 // File generated from our OpenAPI spec by Stainless. See CONTRIBUTING.md for details.
 
 import { APIResource } from '../../../resource';
+import { isRequestOptions } from '../../../core';
 import * as Core from '../../../core';
 import * as LoadBalancersAPI from '../load-balancers';
 import * as HealthAPI from './health';
@@ -37,7 +38,7 @@ export class Pools extends APIResource {
    * ```
    */
   create(params: PoolCreateParams, options?: Core.RequestOptions): Core.APIPromise<Pool> {
-    const { account_id, ...body } = params;
+    const { account_id = this._client.accountId, ...body } = params;
     return (
       this._client.post(`/accounts/${account_id}/load_balancers/pools`, {
         body,
@@ -62,7 +63,7 @@ export class Pools extends APIResource {
    * ```
    */
   update(poolId: string, params: PoolUpdateParams, options?: Core.RequestOptions): Core.APIPromise<Pool> {
-    const { account_id, ...body } = params;
+    const { account_id = this._client.accountId, ...body } = params;
     return (
       this._client.put(`/accounts/${account_id}/load_balancers/pools/${poolId}`, {
         body,
@@ -84,8 +85,16 @@ export class Pools extends APIResource {
    * }
    * ```
    */
-  list(params: PoolListParams, options?: Core.RequestOptions): Core.PagePromise<PoolsSinglePage, Pool> {
-    const { account_id, ...query } = params;
+  list(params?: PoolListParams, options?: Core.RequestOptions): Core.PagePromise<PoolsSinglePage, Pool>;
+  list(options?: Core.RequestOptions): Core.PagePromise<PoolsSinglePage, Pool>;
+  list(
+    params: PoolListParams | Core.RequestOptions = {},
+    options?: Core.RequestOptions,
+  ): Core.PagePromise<PoolsSinglePage, Pool> {
+    if (isRequestOptions(params)) {
+      return this.list({}, params);
+    }
+    const { account_id = this._client.accountId, ...query } = params;
     return this._client.getAPIList(`/accounts/${account_id}/load_balancers/pools`, PoolsSinglePage, {
       query,
       ...options,
@@ -105,10 +114,19 @@ export class Pools extends APIResource {
    */
   delete(
     poolId: string,
-    params: PoolDeleteParams,
+    params?: PoolDeleteParams,
+    options?: Core.RequestOptions,
+  ): Core.APIPromise<PoolDeleteResponse>;
+  delete(poolId: string, options?: Core.RequestOptions): Core.APIPromise<PoolDeleteResponse>;
+  delete(
+    poolId: string,
+    params: PoolDeleteParams | Core.RequestOptions = {},
     options?: Core.RequestOptions,
   ): Core.APIPromise<PoolDeleteResponse> {
-    const { account_id } = params;
+    if (isRequestOptions(params)) {
+      return this.delete(poolId, {}, params);
+    }
+    const { account_id = this._client.accountId } = params;
     return (
       this._client.delete(
         `/accounts/${account_id}/load_balancers/pools/${poolId}`,
@@ -134,10 +152,18 @@ export class Pools extends APIResource {
    * ```
    */
   bulkEdit(
-    params: PoolBulkEditParams,
+    params?: PoolBulkEditParams,
+    options?: Core.RequestOptions,
+  ): Core.PagePromise<PoolsSinglePage, Pool>;
+  bulkEdit(options?: Core.RequestOptions): Core.PagePromise<PoolsSinglePage, Pool>;
+  bulkEdit(
+    params: PoolBulkEditParams | Core.RequestOptions = {},
     options?: Core.RequestOptions,
   ): Core.PagePromise<PoolsSinglePage, Pool> {
-    const { account_id, ...body } = params;
+    if (isRequestOptions(params)) {
+      return this.bulkEdit({}, params);
+    }
+    const { account_id = this._client.accountId, ...body } = params;
     return this._client.getAPIList(`/accounts/${account_id}/load_balancers/pools`, PoolsSinglePage, {
       body,
       method: 'patch',
@@ -157,7 +183,7 @@ export class Pools extends APIResource {
    * ```
    */
   edit(poolId: string, params: PoolEditParams, options?: Core.RequestOptions): Core.APIPromise<Pool> {
-    const { account_id, ...body } = params;
+    const { account_id = this._client.accountId, ...body } = params;
     return (
       this._client.patch(`/accounts/${account_id}/load_balancers/pools/${poolId}`, {
         body,
@@ -177,8 +203,17 @@ export class Pools extends APIResource {
    * );
    * ```
    */
-  get(poolId: string, params: PoolGetParams, options?: Core.RequestOptions): Core.APIPromise<Pool> {
-    const { account_id } = params;
+  get(poolId: string, params?: PoolGetParams, options?: Core.RequestOptions): Core.APIPromise<Pool>;
+  get(poolId: string, options?: Core.RequestOptions): Core.APIPromise<Pool>;
+  get(
+    poolId: string,
+    params: PoolGetParams | Core.RequestOptions = {},
+    options?: Core.RequestOptions,
+  ): Core.APIPromise<Pool> {
+    if (isRequestOptions(params)) {
+      return this.get(poolId, {}, params);
+    }
+    const { account_id = this._client.accountId } = params;
     return (
       this._client.get(`/accounts/${account_id}/load_balancers/pools/${poolId}`, options) as Core.APIPromise<{
         result: Pool;
@@ -303,7 +338,7 @@ export interface PoolCreateParams {
   /**
    * Path param: Identifier.
    */
-  account_id: string;
+  account_id?: string;
 
   /**
    * Body param: A short name (tag) for the pool. Only alphanumeric characters,
@@ -392,7 +427,7 @@ export interface PoolUpdateParams {
   /**
    * Path param: Identifier.
    */
-  account_id: string;
+  account_id?: string;
 
   /**
    * Body param: A short name (tag) for the pool. Only alphanumeric characters,
@@ -487,7 +522,7 @@ export interface PoolListParams {
   /**
    * Path param: Identifier.
    */
-  account_id: string;
+  account_id?: string;
 
   /**
    * Query param: The ID of the Monitor to use for checking the health of origins
@@ -500,14 +535,14 @@ export interface PoolDeleteParams {
   /**
    * Identifier.
    */
-  account_id: string;
+  account_id?: string;
 }
 
 export interface PoolBulkEditParams {
   /**
    * Path param: Identifier.
    */
-  account_id: string;
+  account_id?: string;
 
   /**
    * Body param: The email address to send health status notifications to. This field
@@ -521,7 +556,7 @@ export interface PoolEditParams {
   /**
    * Path param: Identifier.
    */
-  account_id: string;
+  account_id?: string;
 
   /**
    * Body param: A list of regions from which to run health checks. Null means every
@@ -616,7 +651,7 @@ export interface PoolGetParams {
   /**
    * Identifier.
    */
-  account_id: string;
+  account_id?: string;
 }
 
 Pools.PoolsSinglePage = PoolsSinglePage;
